Refactor MenuEvents and backend Search API for code health
